+++ satsim/satsim.py
@@ -578,6 +578,7 @@ def image_generator(ssp, output_dir='.', output_debug=False, dir_debug='./Debug'
         ts_collect_start = time.utc_from_list(tt)
         ts_frame_end = time.utc_from_list(tt, t_exposure)
         ts_collect_end = time.utc_from_list(tt, t_frame * num_frames)
+        track_apparent = True
 
         # site
         site_mode = None
@@ -616,6 +617,7 @@ def image_generator(ssp, output_dir='.', output_debug=False, dir_debug='./Debug'
             astrometrics['object'] = pydash.objects.get(ssp, 'geometry.site.track.name', '')
             track_az = pydash.objects.get(ssp, 'geometry.site.track.az', 0)
             track_el = pydash.objects.get(ssp, 'geometry.site.track.el', 0)
+            track_apparent = pydash.objects.get(ssp, 'geometry.site.track.track_apparent', True)
 
             if type(track_az) is not list:
                 track_az = [track_az, track_az]
@@ -631,13 +633,30 @@ def image_generator(ssp, output_dir='.', output_debug=False, dir_debug='./Debug'
                 ts_frame_end = time.utc_from_list(tt, t_exposure)
                 ts_collect_end = time.utc_from_list(tt, t_frame * num_frames)
 
-            star_ra, star_dec, star_tran_os, star_rot_rate = _calculate_star_position_and_motion(ssp, astrometrics,
-                                                                                                 ts_collect_start, ts_collect_end, ts_collect_start, ts_frame_end, t_exposure,
-                                                                                                 h_fpa_pad_os, w_fpa_pad_os,
-                                                                                                 y_fov_pad, x_fov_pad,
-                                                                                                 y_fov, x_fov,
-                                                                                                 y_ifov, x_ifov,
-                                                                                                 observer, track, star_rot, astrometrics['track_mode'], track_az, track_el)
+            star_ra, star_dec, star_tran_os, star_rot_rate = _calculate_star_position_and_motion(
+                ssp,
+                astrometrics,
+                ts_collect_start,
+                ts_collect_end,
+                ts_collect_start,
+                ts_frame_end,
+                t_exposure,
+                h_fpa_pad_os,
+                w_fpa_pad_os,
+                y_fov_pad,
+                x_fov_pad,
+                y_fov,
+                x_fov,
+                y_ifov,
+                x_ifov,
+                observer,
+                track,
+                star_rot,
+                astrometrics['track_mode'],
+                track_az,
+                track_el,
+                track_apparent,
+            )
         else:
             observer = None
             track = None
@@ -755,7 +774,7 @@ def image_generator(ssp, output_dir='.', output_debug=False, dir_debug='./Debug'
                                                                                     h_fpa_pad_os, w_fpa_pad_os,
                                                                                     y_fov, x_fov,
                                                                                     y_to_pix, x_to_pix,
-                                                                                    star_rot, astrometrics['track_mode'])
+                                                                                    star_rot, astrometrics['track_mode'], track_apparent)
             logger.debug('Number of objects {}.'.format(len(obs_os_pix)))
 
             if track_mode is not None:
@@ -766,7 +785,7 @@ def image_generator(ssp, output_dir='.', output_debug=False, dir_debug='./Debug'
                                                                                                          y_fov_pad, x_fov_pad,
                                                                                                          y_fov, x_fov,
                                                                                                          y_ifov, x_ifov,
-                                                                                                         observer, track, star_rot, astrometrics['track_mode'], track_az, track_el)
+                                                                                                         observer, track, star_rot, astrometrics['track_mode'], track_az, track_el, track_apparent)
 
             # if image rendering is disabled, optionally generate analytical observations and return
             if ssp['sim']['mode'] == 'none':
@@ -1215,7 +1234,7 @@ def _gen_objects(ssp, render_mode,
                  h_fpa_pad_os, w_fpa_pad_os,
                  y_fov, x_fov,
                  y_to_pix, x_to_pix,
-                 star_rot, track_mode):
+                 star_rot, track_mode, track_apparent=True):
     """Generate object pixels. TODO move to submodule
     """
     obs_model = []
@@ -1233,6 +1252,9 @@ def _gen_objects(ssp, render_mode,
     enable_fov_filter = ssp['sim'].get('enable_fov_filter', False)
     fov_filter_radius = ssp['sim'].get('fov_filter_radius')
 
+    track_deflection = enable_deflection if track_apparent is True else False
+    track_aberration = enable_light_transit if track_apparent is True else False
+
     az_arr = None
     el_arr = None
     ra_c = None
@@ -1247,8 +1269,8 @@ def _gen_objects(ssp, render_mode,
                 az_arr[1],
                 el_arr[1],
                 ts_mid,
-                deflection=enable_deflection,
-                aberration=enable_light_transit,
+                deflection=track_deflection,
+                aberration=track_aberration,
                 stellar_aberration=enable_stellar_aberration,
             )
         else:
@@ -1256,8 +1278,8 @@ def _gen_objects(ssp, render_mode,
                 observer,
                 track,
                 ts_mid,
-                deflection=enable_deflection,
-                aberration=enable_light_transit,
+                deflection=track_deflection,
+                aberration=track_aberration,
                 stellar_aberration=enable_stellar_aberration,
             )
 
@@ -1400,8 +1422,8 @@ def _gen_objects(ssp, render_mode,
                             az_arr[1],
                             el_arr[1],
                             ts_mid,
-                            deflection=enable_deflection,
-                            aberration=enable_light_transit,
+                            deflection=track_deflection,
+                            aberration=track_aberration,
                             stellar_aberration=enable_stellar_aberration,
                         )
                     else:
@@ -1409,8 +1431,8 @@ def _gen_objects(ssp, render_mode,
                             observer,
                             track,
                             ts_mid,
-                            deflection=enable_deflection,
-                            aberration=enable_light_transit,
+                            deflection=track_deflection,
+                            aberration=track_aberration,
                             stellar_aberration=enable_stellar_aberration,
                         )
 
@@ -1451,6 +1473,7 @@ def _gen_objects(ssp, render_mode,
                         el=el_arr,
                         deflection=enable_deflection,
                         aberration=enable_light_transit,
+                        track_apparent=track_apparent,  # enable apparent correction in pointing
                         stellar_aberration=False,  # disable stellar aberration for target
                     )
                 num_tracks += 1
@@ -1597,11 +1620,16 @@ def _calculate_star_position_and_motion(ssp, astrometrics,
                                         y_fov_pad, x_fov_pad,
                                         y_fov, x_fov,
                                         y_ifov, x_ifov,
-                                        observer, track, star_rot, track_mode, track_az, track_el):
+                                        observer, track, star_rot, track_mode, track_az, track_el, track_apparent=True):
     az, el = _calculate_az_el(ts_collect_start, ts_collect_end, [t_start, t_end], track_az, track_el)
     enable_deflection = ssp['sim']['enable_deflection']
     enable_light_transit = ssp['sim']['enable_light_transit']
     enable_stellar_aberration = ssp['sim']['enable_stellar_aberration']
+
+    if not track_apparent:
+        enable_deflection = False
+        enable_light_transit = False
+
     if track_mode == 'rate':
         track_target = [track]
         star_ra0, star_dec0, _, _, _, _ = get_los(observer, track, t_start, deflection=enable_deflection, aberration=enable_light_transit, stellar_aberration=enable_stellar_aberration)
@@ -1643,6 +1671,7 @@ def _calculate_star_position_and_motion(ssp, astrometrics,
         el=el,
         deflection=enable_deflection,
         aberration=enable_light_transit,
+        track_apparent=track_apparent,
         stellar_aberration=False,
     )
     star_tran_os = [-drr, -dcc]  # stars move in the opposite direction of target